+++ tests/testthat/test-settings.R
@@ -38,3 +38,12 @@ test_that("it uses config settings in same directory if provided", {
 
   expect_equal(settings$exclude, "test")
 })
+
+test_that("it errors if the config file does not end in a newline", {
+
+  f <- tempfile()
+  cat("linters: with_defaults(closed_curly_linter = NULL)", file = f)
+  withr::with_options(c("lintr.linter_file" = f),
+    expect_error(read_settings("foo"), "Malformed config file")
+  )
+})